requirements.txt
aiohttp
async-timeout
attrs
chardet
setuptools
multidict
pillow
pyyaml
typing-extensions
yarl
python-dotenv